backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.32.1
jinja2==3.1.5
python-dotenv==1.0.1
psycopg2-binary==2.9.10
bcrypt==4.2.1
itsdangerous==2.2.0
ruff==0.6.8
python-multipart==0.0.12
pytest==8.3.5

